frontend: remove border-radius from <StartStopButton>, move to a separate folder
