+++ frontend/src/App.css
@@ -32,7 +32,3 @@
   padding: 0 12px;
   flex-grow: 1;
 }
-
-.footer {
-  width: 100vw;
-}
