+++ frontend/src/components/StartStopButton/index.css
@@ -1,0 +1,3 @@
+.start-stop-btn {
+  border-radius: 0 !important;
+}
